backend/pyproject.toml
[project]
name = "openOii-backend"
version = "0.1.0"
description = "openOii AI 漫剧生成平台后端 (FastAPI + SQLModel)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.2.0",
  "sqlmodel>=0.0.31",
  "sqlalchemy>=2.0.30",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.29.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "anthropic>=0.40.0",
  "pillow>=10.0.0",
]

[project.optional-dependencies]
agents = [
  "claude-agent-sdk>=0.1.0",
  "anthropic>=0.40.0",
]

[dependency-groups]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
